backend/requirements.txt
fastapi
uvicorn
playwright
pandas
python-multipart
sse-starlette
openai
python-dotenv
httpx
